server/DelegateServer/requirements.txt
fastapi
azure-storage-blob
uvicorn
pymongo